Created model club and GET & POST routes for club
Created get clubs/ & get clubs/shorthand for clubs that searches either for
all clubs, or a single one if a parameter is given.

Creaated a POST clubs/ that validates the model before saving.
First time using async, and it makes the code much more readable and
shortens it as well.