Fixed bat routing, post /bout & update bout model
Bouts router had a bad path, POST */bouts/bout instead of POST */bouts
Changed the post function to use async/await instead of chaining
promises. Was having issue with chaining and async/await also simpified
the code alot.

Also updated the bout model to keep a string reference of the shorthand
of opponent club, instead of relying on the club stored in opponent
since that is subjected to change when athletes change clubs.